refactor tests. this class is too large and contains multiple testclasses
